ml-backend/tfserving/pyproject.toml
[tool.poetry]
name = "tfserving-demo"
version = "0.1.0"
description = ""
authors = ["Henry Ruiz  <henryruiz22@gmail.com>"]
readme = "README.md"
packages = [{include = "tfserving_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.6.2"
numpy = "^1.23.4"
pillow = "^9.3.0"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
